+++ properties/Connection/CanDropUnknownDatabaseIfNotExists.php
@@ -3,14 +3,20 @@ declare(strict_types = 1);
 
 namespace Properties\Formal\AccessLayer\Connection;
 
-use Formal\AccessLayer\Query;
+use Formal\AccessLayer\{
+    Query,
+    Connection,
+};
 use Fixtures\Formal\AccessLayer\Table\Name;
 use Innmind\BlackBox\{
     Property,
     Set,
+    Runner\Assert,
 };
-use PHPUnit\Framework\Assert;
 
+/**
+ * @implements Property<Connection>
+ */
 final class CanDropUnknownDatabaseIfNotExists implements Property
 {
     private $name;
@@ -22,15 +28,7 @@ final class CanDropUnknownDatabaseIfNotExists implements Property
 
     public static function any(): Set
     {
-        return Set\Property::of(
-            self::class,
-            Name::any(),
-        );
-    }
-
-    public function name(): string
-    {
-        return 'Can drop unknown database if not exists';
+        return Name::any()->map(static fn($name) => new self($name));
     }
 
     public function applicableTo(object $connection): bool
@@ -38,11 +36,11 @@ final class CanDropUnknownDatabaseIfNotExists implements Property
         return true;
     }
 
-    public function ensureHeldBy(object $connection): object
+    public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $rows = $connection(Query\DropTable::ifExists($this->name));
 
-        Assert::assertCount(0, $rows);
+        $assert->count(0, $rows);
 
         return $connection;
     }