+++ properties/Connection/ContentInsertedAfterStartOfTransactionIsAccessible.php
@@ -3,18 +3,22 @@ declare(strict_types = 1);
 
 namespace Properties\Formal\AccessLayer\Connection;
 
-use Formal\AccessLayer\Query\{
-    SQL,
-    Parameter,
-    StartTransaction,
-    Commit,
+use Formal\AccessLayer\{
+    Query\SQL,
+    Query\Parameter,
+    Query\StartTransaction,
+    Query\Commit,
+    Connection,
 };
 use Innmind\BlackBox\{
     Property,
     Set,
+    Runner\Assert,
 };
-use PHPUnit\Framework\Assert;
 
+/**
+ * @implements Property<Connection>
+ */
 final class ContentInsertedAfterStartOfTransactionIsAccessible implements Property
 {
     private string $uuid;
@@ -30,17 +34,12 @@ final class ContentInsertedAfterStartOfTransactionIsAccessible implements Proper
 
     public static function any(): Set
     {
-        return Set\Property::of(
-            self::class,
+        return Set\Composite::immutable(
+            static fn(...$args) => new self(...$args),
             Set\Uuid::any(),
             Set\Strings::madeOf(Set\Chars::ascii())->between(0, 255),
             Set\Integers::any(),
         );
-    }
-
-    public function name(): string
-    {
-        return 'Content inserted after start of a transaction is accessible to query';
     }
 
     public function applicableTo(object $connection): bool
@@ -48,7 +47,7 @@ final class ContentInsertedAfterStartOfTransactionIsAccessible implements Proper
         return true;
     }
 
-    public function ensureHeldBy(object $connection): object
+    public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $connection(new StartTransaction);
 
@@ -60,8 +59,8 @@ final class ContentInsertedAfterStartOfTransactionIsAccessible implements Proper
 
         $rows = $connection(SQL::of("SELECT * FROM `test` WHERE `id` = '{$this->uuid}'"));
 
-        Assert::assertCount(1, $rows);
-        Assert::assertSame(
+        $assert->count(1, $rows);
+        $assert->same(
             $this->uuid,
             $rows
                 ->first()
@@ -71,7 +70,7 @@ final class ContentInsertedAfterStartOfTransactionIsAccessible implements Proper
                     static fn() => null,
                 ),
         );
-        Assert::assertSame(
+        $assert->same(
             $this->username,
             $rows
                 ->first()
@@ -81,7 +80,7 @@ final class ContentInsertedAfterStartOfTransactionIsAccessible implements Proper
                     static fn() => null,
                 ),
         );
-        Assert::assertSame(
+        $assert->same(
             $this->number,
             $rows
                 ->first()