+++ properties/Connection/Insert.php
@@ -8,13 +8,17 @@ use Formal\AccessLayer\{
     Query,
     Table,
     Row,
+    Connection,
 };
 use Innmind\BlackBox\{
     Property,
     Set,
+    Runner\Assert,
 };
-use PHPUnit\Framework\Assert;
 
+/**
+ * @implements Property<Connection>
+ */
 final class Insert implements Property
 {
     private string $uuid;
@@ -26,15 +30,7 @@ final class Insert implements Property
 
     public static function any(): Set
     {
-        return Set\Property::of(
-            self::class,
-            Set\Uuid::any(),
-        );
-    }
-
-    public function name(): string
-    {
-        return 'Insert';
+        return Set\Uuid::any()->map(static fn($uuid) => new self($uuid));
     }
 
     public function applicableTo(object $connection): bool
@@ -42,12 +38,12 @@ final class Insert implements Property
         return true;
     }
 
-    public function ensureHeldBy(object $connection): object
+    public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $select = SQL::of("SELECT * FROM `test` WHERE `id` = '{$this->uuid}'");
         $rows = $connection($select);
 
-        Assert::assertCount(0, $rows);
+        $assert->count(0, $rows);
 
         $sequence = $connection(Query\Insert::into(
             new Table\Name('test'),
@@ -58,12 +54,12 @@ final class Insert implements Property
             ]),
         ));
 
-        Assert::assertCount(0, $sequence);
+        $assert->count(0, $sequence);
 
         $rows = $connection($select);
 
-        Assert::assertCount(1, $rows);
-        Assert::assertSame(
+        $assert->count(1, $rows);
+        $assert->same(
             $this->uuid,
             $rows
                 ->first()
@@ -73,7 +69,7 @@ final class Insert implements Property
                     static fn() => null,
                 ),
         );
-        Assert::assertSame(
+        $assert->same(
             'foo',
             $rows
                 ->first()
@@ -83,7 +79,7 @@ final class Insert implements Property
                     static fn() => null,
                 ),
         );
-        Assert::assertSame(
+        $assert->same(
             42,
             $rows
                 ->first()
@@ -93,7 +89,7 @@ final class Insert implements Property
                     static fn() => null,
                 ),
         );
-        Assert::assertSame(
+        $assert->same(
             [
                 'id' => $this->uuid,
                 'username' => 'foo',