+++ properties/Connection/ParameterTypesCanBeSpecified.php
@@ -3,17 +3,21 @@ declare(strict_types = 1);
 
 namespace Properties\Formal\AccessLayer\Connection;
 
-use Formal\AccessLayer\Query\{
-    SQL,
-    Parameter,
-    Parameter\Type,
+use Formal\AccessLayer\{
+    Query\SQL,
+    Query\Parameter,
+    Query\Parameter\Type,
+    Connection,
 };
 use Innmind\BlackBox\{
     Property,
     Set,
+    Runner\Assert,
 };
-use PHPUnit\Framework\Assert;
 
+/**
+ * @implements Property<Connection>
+ */
 final class ParameterTypesCanBeSpecified implements Property
 {
     private string $uuid;
@@ -29,17 +33,12 @@ final class ParameterTypesCanBeSpecified implements Property
 
     public static function any(): Set
     {
-        return Set\Property::of(
-            self::class,
+        return Set\Composite::immutable(
+            static fn(...$args) => new self(...$args),
             Set\Uuid::any(),
             Set\Strings::madeOf(Set\Chars::ascii())->between(0, 255),
             Set\Integers::any(),
         );
-    }
-
-    public function name(): string
-    {
-        return 'Parameter types can be specified';
     }
 
     public function applicableTo(object $connection): bool
@@ -47,7 +46,7 @@ final class ParameterTypesCanBeSpecified implements Property
         return true;
     }
 
-    public function ensureHeldBy(object $connection): object
+    public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $insert = SQL::of('INSERT INTO `test` VALUES (?, ?, ?);')
             ->with(Parameter::of($this->uuid, Type::string))
@@ -57,8 +56,8 @@ final class ParameterTypesCanBeSpecified implements Property
 
         $rows = $connection(SQL::of("SELECT * FROM `test` WHERE `id` = '{$this->uuid}'"));
 
-        Assert::assertCount(1, $rows);
-        Assert::assertSame(
+        $assert->count(1, $rows);
+        $assert->same(
             $this->uuid,
             $rows
                 ->first()
@@ -68,7 +67,7 @@ final class ParameterTypesCanBeSpecified implements Property
                     static fn() => null,
                 ),
         );
-        Assert::assertSame(
+        $assert->same(
             $this->username,
             $rows
                 ->first()
@@ -78,7 +77,7 @@ final class ParameterTypesCanBeSpecified implements Property
                     static fn() => null,
                 ),
         );
-        Assert::assertSame(
+        $assert->same(
             $this->number,
             $rows
                 ->first()